% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barogram.R
\name{barogram}
\alias{barogram}
\title{Barogram Plot}
\usage{
barogram(data, alt_unit)
}
\arguments{
\item{data}{data frame from the output of the IGCparse() function}

\item{alt_unit}{string of either "feet" or "metre" to indicate what units the altitude values should be}
}
\value{
ggplot2 object
}
\description{
Barogram Plot
}
\examples{
df <- IGCparse("2021-02-05-XCS-AAA-03.igc")
barogram(df,"metre")

}
