% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climb_plot.R
\name{climb_plot}
\alias{climb_plot}
\title{Climb Plot}
\usage{
climb_plot(data, rate_unit)
}
\arguments{
\item{data}{data frame from the output of the IGCparse() function}

\item{rate_unit}{string of either "knots", "m/s" or "feet/min" to indicate units for climb rate}
}
\value{
ggplot2 object
}
\description{
Climb Plot
}
\examples{
df <- IGCparse("2021-02-05-XCS-AAA-03.igc")
climb_plot(df,"knots")


}
