% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGCparse.R
\name{IGCparse}
\alias{IGCparse}
\title{Parse IGC file}
\usage{
IGCparse(filepath)
}
\arguments{
\item{filepath}{filepath of IGC file contained within " "}
}
\value{
A data frame, with time in seconds, latitude and longitude in decimal degrees, pressure
and GNSS altitude in meters, climb rate in m/s
}
\description{
Parse IGC file
}
\examples{
IGCparse("2021-02-05-XCS-AAA-03.igc")


}
