% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tow_release.R
\name{tow_release}
\alias{tow_release}
\title{Tow release height}
\usage{
tow_release(data, alt_unit)
}
\arguments{
\item{data}{data frame from the output of the IGCparse() function}

\item{alt_unit}{string of either "feet" or "metre" to indicate what units the altitude values should be}
}
\value{

}
\description{
Tow release height
}
\examples{


}
